Refact BlogPost
