+++ src/components/BlogRoll.js
@@ -72,7 +72,7 @@ export default () => (
       query BlogRollQuery {
         allMarkdownRemark(
           sort: { order: DESC, fields: [frontmatter___date] }
-          filter: { frontmatter: { templateKey: { eq: "blog-post" } } }
+          filter: { frontmatter: { templateKey: { eq: "BlogPost" } } }
         ) {
           edges {
             node {
