+++ src/pages/blog/2016-12-17-making-sense-of-the-scaas-new-flavor-wheel.md
@@ -1,5 +1,5 @@
 ---
-templateKey: blog
+templateKey: BlogPost
 title: Making sense of the SCAA’s new Flavor Wheel
 date: 2016-12-17T15:04:10.000Z
 featuredpost: false
