+++ static/admin/config.yml
@@ -12,7 +12,7 @@ media_folder: static/img
 public_folder: /img
 
 collections:
-  - name: "blog"
+  - name: "BlogPost"
     label: "Blog"
     folder: "src/pages/blog"
     create: true
